\name{DES1}
\alias{DES1}
\docType{data}
\title{DES1 \{dispersion-package\}}
\description{This is an .rda file containing a design in a matrix.}
\usage{data(DES1)}
\format{Three columns and eighteen rows.}
\details{This design is presented in Table 2 of [1], referred to as design VI, obtained by using a compound criterion, with 5 degrees of freedom to estimate the pure error and 3 degrees of freedom for estimating lack of fit. It is a completely randomized design for three factors in the spherical region.}
\source{[1] GILMOUR, S. G.; TRINCA, L. A. Optimum design of experiments for statistical inference (with discussion). J. Roy. Statist. Soc., Series C (Applied Statistics), 61,345-401. 2012.}
\keyword{datasets}
